# perf/perftool: Fix unicode decode errors