Perf test sometimes generates an output that can cause this test to
error out. Utf-8 decoding fails when certain invalid characters are encountered.

Fix this issue by ignoring such invalid characters.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>