@sacsant 2022-12-31
Before the fix:

 (1/2) perf/perftool.py:Perftool.test_perf_test: STARTED
 (1/2) perf/perftool.py:Perftool.test_perf_test: FAIL: 6 Test failed (100.04 s)
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite: STARTED
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite: ERROR: 'utf-8' codec can't decode byte 0x80 in position 3689: invalid start byte (250.59 s)

After the fix:

 (1/2) perf/perftool.py:Perftool.test_perf_test: STARTED
 (1/2) perf/perftool.py:Perftool.test_perf_test: FAIL: 7 Test failed (60.01 s)
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite: STARTED
 (2/2) perf/perftool.py:Perftool.test_perf_testsuite: FAIL: 22 Test failed (249.60 s)